refactor : 변환, 검증 객체 패키지 위치 변경
