+++ src/main/java/lotto/view/converter/validator/LottoInputValidator.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.view.converter.validator;
 
 public class LottoInputValidator {
