+++ src/main/java/lotto/view/converter/validator/LottoMoneyValidator.java
@@ -1,4 +1,4 @@
-package lotto.service;
+package lotto.view.converter.validator;
 
 import lotto.constant.LottoConstant;
