+++ src/test/java/lotto/service/LottoMoneyValidatorTest.java
@@ -2,6 +2,7 @@ package lotto.service;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import lotto.view.converter.validator.LottoMoneyValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
